Fix integer overflow in TFL::FullyConnectedOp::verify() by using int64_t to store num_elements.
Consider the following tfl.fullyconnectd_op, the numerics in the verify function suffer from a integer overflow and produce undesired TFL Converter errors.

```
"tfl.fully_connected"(...) <{...}> :
  (tensor<2048x128xf32>, tensor<1049088x128xf32>, none) -> tensor<2048x1049088xf32>
```

Produces the error- `'tfl.fully_connected' op expect 'output' num_elements % 1049088 == 0, got 'tensor<2048x1049088xf32>'`, because the `num_elements` is calculated to be 2,148,532,224 which is slightly larger than the maximum value representable by a standard 32-bit signed integer (int32_t), which is 2,147,483,647. So using int64_t in the code.

PiperOrigin-RevId: 747981025

+++ tensorflow/compiler/mlir/lite/ir/tfl_ops.cc
@@ -1529,7 +1529,7 @@ LogicalResult FullyConnectedOp::verify() {
 
   // Input's element size must be multiple of parameter's z_in dimension.
   const int z_in = filter_type.getDimSize(1);
-  const int num_input_elements = input_type.getNumElements();
+  const int64_t num_input_elements = input_type.getNumElements();
   if (z_in != 0 && num_input_elements % z_in != 0) {
     return op.emitOpError(llvm::formatv(
                "expect 'input' num_elements % {0} == 0, got input type ", z_in))
@@ -1545,7 +1545,7 @@ LogicalResult FullyConnectedOp::verify() {
       return mlir::success();
     }
 
-    const int num_output_elements = output_type.getNumElements();
+    const int64_t num_output_elements = output_type.getNumElements();
     const int z_out = filter_type.getDimSize(0);
     if (num_output_elements % z_out != 0) {
       return op.emitOpError(llvm::formatv(
